:recycle: Refactor: 공통 컴포넌트 core 모듈로 이동
